IoT/requirements.txt
numpy>=1.13.0
# opencv-python>=3.2.0.7
pandas>=0.19.2
paho-mqtt
pillow>=6.2.0
plotly>=2.0.9
python-dateutil>=2.6.0
requests>=2.20.0
scikit-image>=0.13.0
scikit-learn>=0.18.1
scipy>=0.19.0
sk-video>=1.1.8
sklearn
PyQt5==5.10.1
tensorflow-gpu
keras
kito
